# swaggerDoc.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Minha Feira API - Manual",
    "description": "API para gerenciamento de feiras. \nLinks úteis:\n- [Minha Feira Repositório](https://github.com/FeirasProjeto)",
    "contact": {
      "email": "feirass.projeto@gmail.com"
    },
    "version": "1.0.11"
  },
  "externalDocs": {
    "description": "Descubra mais sobre a aplicação Minha Feira",
    "url": "https://github.com/FeirasProjeto"
  },
  "servers": [
    {
      "url": "https://website-feira.vercel.app"
    }
  ],
  "paths": {
    "/user": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Lista todos os usuários",
        "description": "Retorna uma lista de todos os usuários cadastrados no sistema.",
        "operationId": "getAllUsers",
        "responses": {
          "200": {
            "description": "Lista de usuários encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Cria um novo usuário",
        "description": "Adiciona um novo usuário ao sistema",
        "operationId": "createUser",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                "nome": {
                  "type": "string",
                  "example": "João Silva"
                },
                "email": {
            "type": "string",
            "example": "joao.silva@email.com"
          },
          "senha": {
          "type": "string",
          "example": "Senha123!"
            
          },
          "telefone": {
            "type": "string",
            "example": "+55 11 91234-5678"
          },
          "feirante": {
            "type": "boolean"
          }
                
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuário criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string",
                    "example": "07731e94-df84-4f00-98ac-972c469bdf66"
                  },
                "nome": {
                  "type": "string",
                  "example": "João Silva"
                },
                "email": {
            "type": "string",
            "example": "joao.silva@email.com"
          },
          "senha": {
          "type": "string",
          "example": "$2b$12$y4TTJih/asTQGJEBSBJlo.yP6Hv/.Ci5z1bTWUf3Mm1an10aUe16m"
            
          },
          "telefone": {
            "type": "string",
            "example": "+55 11 91234-5678"
          },
          "feirante": {
            "type": "boolean"
          }
                
                }
                }
              }
            }
          },
          "400": {
            "description": "Entrada inválida ou e-mail já cadastrado"
          }
        }
      }
    },
    "/user/{id}": {
      "put": {
        "tags": [
          "User"
        ],
        "summary": "Atualiza um usuário existente",
        "description": "Atualiza informações de um usuário cadastrado pelo seu ID",
        "operationId": "updateUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "feirante": {
                    "type": "boolean",
                    "example": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuário atualizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "404": {
            "description": "Usuário não encontrado"
          }
        }
      },
      "delete": {
        "tags": [
          "User"
        ],
        "summary": "Remove um usuário pelo ID",
        "description": "Remove um usuário específico",
        "operationId": "deleteUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário removido com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "404": {
            "description": "Usuário não encontrado"
          }
        }
      }
    },
    "/favoritos": {
      "get": {
        "tags": [
          "Favorito"
        ],
        "summary": "Lista todos os favoritos",
        "description": "Retorna uma lista de todos os itens favoritados",
        "operationId": "getFavoritos",
        "responses": {
          "200": {
            "description": "Operação bem-sucedida",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Favorito"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Nenhum favorito encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Nenhum favorito encontrado."
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Favorito"
        ],
        "summary": "Cria um novo favorito",
        "description": "Adiciona um novo item à lista de favoritos",
        "operationId": "createFavorito",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": {
                    "type": "string",
                    "example": "123"
                  },
                  "feiraId": {
                    "type": "string",
                    "example": "456"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Favorito criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Favorito"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos fornecidos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Dados inválidos fornecidos."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/favoritos/{id}": {
      "delete": {
        "tags": [
          "Favorito"
        ],
        "summary": "Remove um favorito pelo ID",
        "description": "Remove um favorito específico",
        "operationId": "deleteFavorito",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Favorito removido",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Favorito removido com sucesso."
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Favorito não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Favorito não encontrado."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/feira": {
      "get": {
        "tags": [
          "Feira"
        ],
        "summary": "Lista todas as feiras",
        "description": "Retorna a lista de todas as feiras cadastradas",
        "operationId": "getFeiras",
        "responses": {
          "200": {
            "description": "Lista de feiras encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Feira"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Feira"
        ],
        "summary": "Adiciona uma nova feira ao sistema",
        "description": "Cria uma nova feira e insere no banco de dados",
        "operationId": "addFeira",
        "requestBody": {
          "description": "Dados da nova feira",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Feira"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Feira criada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Feira"
                }
              }
            }
          },
          "400": {
            "description": "Entrada inválida"
          },
          "422": {
            "description": "Exceção de validação"
          }
        }
      }
    },
    "/feira/{feiraId}": {
      "put": {
        "tags": [
          "Feira"
        ],
        "summary": "Atualiza feira por ID",
        "description": "Atualiza os detalhes de uma feira específica pelo ID",
        "operationId": "updateFeiraById",
        "parameters": [
          {
            "name": "feiraId",
            "in": "path",
            "description": "ID da feira",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nome": {
                    "type": "string",
                    "example": "Feira da Praça"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Operação bem-sucedida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Feira"
                }
              }
            }
          },
          "400": {
            "description": "ID inválido fornecido"
          },
          "404": {
            "description": "Feira não encontrada"
          }
        }
      },
      "delete": {
        "tags": [
          "Feira"
        ],
        "summary": "Deleta feira por ID",
        "description": "Remove uma feira específica pelo ID",
        "operationId": "deleteFeiraById",
        "parameters": [
          {
            "name": "feiraId",
            "in": "path",
            "description": "ID da feira a ser deletada",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operação bem-sucedida",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Feira deletada com sucesso"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "ID inválido fornecido"
          },
          "404": {
            "description": "Feira não encontrada"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "c1a45f50-f8a4-4dc1-b9f3-d3a7e556a813"
          },
          "nome": {
            "type": "string",
            "example": "João Silva"
          },
          "email": {
            "type": "string",
            "example": "joao.silva@email.com"
          },
          "telefone": {
            "type": "string",
            "example": "+55 11 91234-5678"
          },
          "feirante": {
            "type": "boolean"
          },
          "_count": {
            "type": "object",
            "properties": {
            "favoritos": {
              "type": "number",
              "example": 2
            }
          }
        }
          }
        },
      "Favorito": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "f1c35b00-bf9e-44dc-bb6a-4c017d56e715"
          },
          "userId": {
            "type": "string",
            "example": "c1a45f50-f8a4-4dc1-b9f3-d3a7e556a813"
          },
          "feiraId": {
            "type": "string",
            "example": "10"
          },
          "feira": {
            "$ref": "#/components/schemas/Feira"
          }
        }
      },
      "Feira": {
        "type": "object",
        "properties": {
          "nome": {
            "type": "string",
            "example": "Feira Orgânica Centro"
          },
          "endereco": {
            "type": "string",
            "example": "Rua das Flores"
          },
          "numero": {
            "type": "string",
            "example": "123"
          },
          "cidade": {
            "type": "string",
            "example": "São Paulo"
          },
          "coordenada": {
            "type": "string",
            "example": "-31.7695952071496, -52.34144075544093"
          },
          "horario": {
            "type": "string",
            "example": "09:00 - 16:00"
          },
          "data": {
            "type": "string",
            "format": "date",
            "example": "2024-11-01"
          },
          "descricao": {
            "type": "string",
            "example": "Feira com produtos orgânicos."
          },
          "imagem": {
            "type": "string",
            "example": "http://example.com/imagem.jpg"
          },
          "diaSemana": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DiaSemana"
            }
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Tag"
            }
          },
          "userId": {
            "type": "string",
            "example": "c1a45f50-f8a4-4dc1-b9f3-d3a7e556a813"
            
          }
        }
      },
      "DiaSemana": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "d1b0bfa6-5a6f-4efb-8eb5-c8c4c1c5ee5a"
          }
        }
      },
      "Tag": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "t2a1b8d5-1f7f-4c51-83c0-4a8db24ab70d"
          }
        }
      }
    }
  }
}